requirements.txt
torch
transformers
tensorflow
keras
pandas
numpy
scikit-learn
matplotlib
opencv-python
requests
pillow
